+++ propulsion-archiver/Infrastructure.fs
@@ -27,11 +27,22 @@ type LoggerConfigurationExtensions() =
         c.MinimumLevel.Override("Microsoft.Azure.Documents.ChangeFeedProcessor", cfpl)
         |> fun c -> if verbose then c else c.ExcludeChangeFeedProcessorV2InternalDiagnostics()
 
+    [<System.Runtime.CompilerServices.Extension>]
+    static member inline ForwardMetricsFromEquinoxAndPropulsionToPrometheus(a : Configuration.LoggerSinkConfiguration, appName, group, metrics) =
+        let customTags = ["app", appName]
+        a.Logger(fun l ->
+            l.WriteTo.Sink(Equinox.CosmosStore.Core.Log.InternalMetrics.Stats.LogSink())
+              |> fun l -> if not metrics then l else
+                            l.WriteTo.Sink(Equinox.CosmosStore.Prometheus.LogSink(customTags))
+                             .WriteTo.Sink(Propulsion.Prometheus.LogSink(customTags, group))
+                             .WriteTo.Sink(Propulsion.CosmosStore.Prometheus.LogSink(customTags))
+              |> ignore) |> ignore
+
 [<System.Runtime.CompilerServices.Extension>]
 type Logging() =
 
     [<System.Runtime.CompilerServices.Extension>]
-    static member Configure(configuration : LoggerConfiguration, verbose, (logSyncToConsole, minRu), cfpVerbose) =
+    static member Configure(configuration : LoggerConfiguration, appName, group, verbose, (logSyncToConsole, minRu), cfpVerbose, metrics) =
         configuration
             .Destructure.FSharpTypes()
             .Enrich.FromLogContext()
@@ -40,21 +51,22 @@ type Logging() =
         |> fun c -> let ingesterLevel = if logSyncToConsole then Events.LogEventLevel.Debug else Events.LogEventLevel.Information
                     c.MinimumLevel.Override(typeof<Propulsion.Streams.Scheduling.StreamSchedulingEngine>.FullName, ingesterLevel)
         |> fun c -> let generalLevel = if verbose then Events.LogEventLevel.Information else Events.LogEventLevel.Warning
-                    c.MinimumLevel.Override(typeof<Propulsion.Cosmos.Internal.Writer.Result>.FullName, generalLevel)
+                    c.MinimumLevel.Override(typeof<Propulsion.CosmosStore.Internal.Writer.Result>.FullName, generalLevel)
         |> fun c ->
             let t = "[{Timestamp:HH:mm:ss} {Level:u3}] {partitionKeyRangeId,2} {Message:lj} {Properties}{NewLine}{Exception}"
             let t = if verbose then t else t.Replace("{Properties}", "")
             let configure (a : Configuration.LoggerSinkConfiguration) : unit =
+                a.ForwardMetricsFromEquinoxAndPropulsionToPrometheus(appName, group, metrics)
                 a.Logger(fun l ->
-                    l.WriteTo.Sink(Equinox.Cosmos.Store.Log.InternalMetrics.Stats.LogSink()) |> ignore) |> ignore
-                a.Logger(fun l ->
-                    let isEqx = Filters.Matching.FromSource<Equinox.Cosmos.Core.Context>().Invoke
-                    let isWriterB = Filters.Matching.FromSource<Propulsion.Cosmos.Internal.Writer.Result>().Invoke
+                    let isEqx = Filters.Matching.FromSource<Equinox.CosmosStore.CosmosStoreContext>().Invoke
+                    let isWriterB = Filters.Matching.FromSource<Propulsion.CosmosStore.Internal.Writer.Result>().Invoke
                     let l = if logSyncToConsole then l else l.Filter.ByExcluding(fun x -> isEqx x || isWriterB x)
                     let isCheaperThan minRu = function
-                        | Equinox.Cosmos.Store.Log.InternalMetrics.Stats.CosmosMetric
-                            (Equinox.Cosmos.Store.Log.Event.SyncSuccess m | Equinox.Cosmos.Store.Log.Event.SyncConflict m) ->
-                                m.ru < minRu
+                        | Equinox.CosmosStore.Core.Log.MetricEvent
+                            (Equinox.CosmosStore.Core.Log.Metric.SyncSuccess m
+                                | Equinox.CosmosStore.Core.Log.Metric.SyncConflict m
+                                | Equinox.CosmosStore.Core.Log.Metric.SyncResync m) ->
+                            m.ru < minRu
                         | _ -> false
                     let l = match minRu with Some mru -> l.Filter.ByExcluding(fun x -> isCheaperThan mru x) | None -> l
                     l.WriteTo.Console(theme=Sinks.SystemConsole.Themes.AnsiConsoleTheme.Code, outputTemplate=t) |> ignore)