+++ propulsion-archiver/Program.fs
@@ -1,7 +1,6 @@
 ﻿module ArchiverTemplate.Program
 
-open Equinox.Cosmos
-open Propulsion.Cosmos
+open Propulsion.CosmosStore
 open Serilog
 open System
 
@@ -31,6 +30,8 @@ module Args =
         | [<AltCommandLine "-S"; Unique>]   SyncVerbose
         | [<AltCommandLine "-t"; Unique>]   RuThreshold of float
         | [<AltCommandLine "-C"; Unique>]   CfpVerbose
+        | [<AltCommandLine "-p"; Unique>]   PrometheusPort of int
+        | [<AltCommandLine "-k"; Unique>]   MaxKib of int
 
         | [<CliPrefix(CliPrefix.None); AltCommandLine "cosmos"; Unique(*ExactlyOnce is not supported*); Last>] SrcCosmos of ParseResults<CosmosSourceParameters>
         interface IArgParserTemplate with
@@ -42,16 +43,21 @@ module Args =
                 | Verbose ->                "request Verbose Logging. Default: off"
                 | SyncVerbose ->            "request Logging for Sync operations (Writes). Default: off"
                 | RuThreshold _ ->          "minimum request charge required to log. Default: 0"
+                | MaxKib _ ->               "max KiB to submit to Sync operation. Default: 512"
                 | CfpVerbose ->             "request Verbose Change Feed Processor Logging. Default: off"
+                | PrometheusPort _ ->       "port from which to expose a Prometheus /metrics endpoint. Default: off"
 
                 | SrcCosmos _ ->            "Cosmos input parameters."
     and Arguments(c : Configuration, a : ParseResults<Parameters>) =
         member val ConsumerGroupName =      a.GetResult ConsumerGroupName
         member val MaxReadAhead =           a.GetResult(MaxReadAhead, 32)
         member val MaxWriters =             a.GetResult(MaxWriters, 4)
+        member val MaxBytes =               a.GetResult(MaxKib, 512) * 1024
         member val Verbose =                a.Contains Parameters.Verbose
         member val CfpVerbose =             a.Contains CfpVerbose
         member val SyncLogging =            a.Contains SyncVerbose, a.TryGetResult RuThreshold
+        member val PrometheusPort =         a.TryGetResult PrometheusPort
+        member val MetricsEnabled =         a.Contains PrometheusPort
         member val StatsInterval =          TimeSpan.FromMinutes 1.
         member val StateInterval =          TimeSpan.FromMinutes 5.
         member val private Source : CosmosSourceArguments =
@@ -60,19 +66,19 @@ module Args =
             | _ -> raise (MissingArg "Must specify cosmos for Src")
         member x.SourceParams() =
             let srcC = x.Source
-            let disco, db =
+            let connection, db =
                 let dstC : CosmosSinkArguments = srcC.Sink
                 match srcC.LeaseContainer, dstC.LeaseContainer with
                 | None, None ->     srcC.Connection, { database = srcC.Database; container = srcC.Container + "-aux" }
                 | Some sc, None ->  srcC.Connection, { database = srcC.Database; container = sc }
                 | None, Some dc ->  dstC.Connection, { database = dstC.Database; container = dc }
                 | Some _, Some _ -> raise (MissingArg "LeaseContainerSource and LeaseContainerDestination are mutually exclusive - can only store in one database")
-            Log.Information("Archiving... {dop} writers, max {maxReadAhead} batches read ahead", x.MaxWriters, x.MaxReadAhead)
+            Log.Information("Archiving... {dop} writers, max {maxReadAhead} batches read ahead, max write batch {maxKib} KiB", x.MaxWriters, x.MaxReadAhead, x.MaxBytes / 1024)
             Log.Information("Monitoring Group {leaseId} in Database {db} Container {container} with maximum document count limited to {maxDocuments}",
                 x.ConsumerGroupName, db.database, db.container, Option.toNullable srcC.MaxDocuments)
             if srcC.FromTail then Log.Warning("(If new projector group) Skipping projection of all existing events.")
             srcC.LagFrequency |> Option.iter<TimeSpan> (fun s -> Log.Information("Dumping lag stats at {lagS:n0}s intervals", s.TotalSeconds))
-            (srcC, (disco, db, x.ConsumerGroupName, srcC.FromTail, srcC.MaxDocuments, srcC.LagFrequency))
+            (srcC, (Equinox.Cosmos.Discovery.FromConnectionString connection, db, x.ConsumerGroupName, srcC.FromTail, srcC.MaxDocuments, srcC.LagFrequency))
     and [<NoEquality; NoComparison>] CosmosSourceParameters =
         | [<AltCommandLine "-Z"; Unique>]   FromTail
         | [<AltCommandLine "-md"; Unique>]  MaxDocuments of int
@@ -105,12 +111,14 @@ module Args =
 
                 | DstCosmos _ ->            "CosmosDb Sink parameters."
     and CosmosSourceArguments(c : Configuration, a : ParseResults<CosmosSourceParameters>) =
+        let discovery =                     a.TryGetResult CosmosSourceParameters.Connection |> Option.defaultWith (fun () -> c.CosmosConnection)
         member val FromTail =               a.Contains CosmosSourceParameters.FromTail
         member val MaxDocuments =           a.TryGetResult MaxDocuments
         member val LagFrequency =           a.TryGetResult LagFreqM |> Option.map TimeSpan.FromMinutes
         member val LeaseContainer =         a.TryGetResult CosmosSourceParameters.LeaseContainer
         member val Mode =                   a.GetResult(CosmosSourceParameters.ConnectionMode, Equinox.Cosmos.ConnectionMode.Direct)
-        member val Connection =             a.TryGetResult CosmosSourceParameters.Connection |> Option.defaultWith (fun () -> c.CosmosConnection) |> Discovery.FromConnectionString
+        member val Connection =             discovery
+        member val Discovery : Equinox.Cosmos.Discovery = Equinox.Cosmos.Discovery.FromConnectionString discovery
         member val Database =               a.TryGetResult CosmosSourceParameters.Database   |> Option.defaultWith (fun () -> c.CosmosDatabase)
         member val Container =              a.GetResult CosmosSourceParameters.Container
         member val Timeout =                a.GetResult(CosmosSourceParameters.Timeout, 5.) |> TimeSpan.FromSeconds
@@ -121,15 +129,15 @@ module Args =
             | Some (DstCosmos cosmos) -> CosmosSinkArguments (c, cosmos)
             | _ -> raise (MissingArg "Must specify cosmos for Sink")
         member x.MonitoringParams() =
-            let Discovery.UriAndKey (endpointUri, _) as discovery = x.Connection
+            let (Equinox.Cosmos.Discovery.UriAndKey (endpointUri, _)) = x.Discovery
             Log.Information("Source CosmosDb {mode} {endpointUri} Database {database} Container {container}",
                 x.Mode, endpointUri, x.Database, x.Container)
             Log.Information("Source CosmosDb timeout {timeout}s; Throttling retries {retries}, max wait {maxRetryWaitTime}s",
                 (let t = x.Timeout in t.TotalSeconds), x.Retries, (let t = x.MaxRetryWaitTime in t.TotalSeconds))
             let c = Equinox.Cosmos.Connector(x.Timeout, x.Retries, x.MaxRetryWaitTime, Log.Logger, mode=x.Mode)
-            c, discovery, { database = x.Database; container = x.Container }
+            c, x.Discovery, { database = x.Database; container = x.Container }
     and [<NoEquality; NoComparison>] CosmosSinkParameters =
-        | [<AltCommandLine "-m">]           ConnectionMode of Equinox.Cosmos.ConnectionMode
+        | [<AltCommandLine "-m">]           ConnectionMode of Microsoft.Azure.Cosmos.ConnectionMode
         | [<AltCommandLine "-s">]           Connection of string
         | [<AltCommandLine "-d">]           Database of string
         | [<AltCommandLine "-c">]           Container of string
@@ -148,23 +156,27 @@ module Args =
                 | Retries _ ->              "specify operation retries. Default: 0."
                 | RetriesWaitTime _ ->      "specify max wait-time for retry when being throttled by Cosmos in seconds. Default: 5."
     and CosmosSinkArguments(c : Configuration, a : ParseResults<CosmosSinkParameters>) =
-        member val Mode =                   a.GetResult(ConnectionMode, Equinox.Cosmos.ConnectionMode.Direct)
-        member val Connection =             a.TryGetResult Connection |> Option.defaultWith (fun () -> c.CosmosConnection) |> Discovery.FromConnectionString
-        member val Database =               a.TryGetResult Database   |> Option.defaultWith (fun () -> c.CosmosDatabase)
-        member val Container =              a.TryGetResult Container  |> Option.defaultWith (fun () -> c.CosmosContainer)
+        let discovery =                     a.TryGetResult Connection |> Option.defaultWith (fun () -> c.CosmosConnection)
+        member val Connection =             discovery
+        member val Mode =                   a.GetResult(ConnectionMode, Microsoft.Azure.Cosmos.ConnectionMode.Direct)
+        member val Discovery =              Equinox.CosmosStore.Discovery.ConnectionString discovery
         member val LeaseContainer =         a.TryGetResult LeaseContainer
         member val Timeout =                a.GetResult(CosmosSinkParameters.Timeout, 5.) |> TimeSpan.FromSeconds
         member val Retries =                a.GetResult(CosmosSinkParameters.Retries, 0)
         member val MaxRetryWaitTime =       a.GetResult(RetriesWaitTime, 5.) |> TimeSpan.FromSeconds
-        /// Connect with the provided parameters and/or environment variables
-        member x.Connect appName : Async<Equinox.Cosmos.Connection> =
-            let Discovery.UriAndKey (endpointUri, _masterKey) as discovery = x.Connection
-            Log.Information("Destination CosmosDb {mode} {endpointUri} Database {database} Container {container}",
-                x.Mode, endpointUri, x.Database, x.Container)
-            Log.Information("Destination CosmosDb timeout {timeout}s; Throttling retries {retries}, max wait {maxRetryWaitTime}s",
-                (let t = x.Timeout in t.TotalSeconds), x.Retries, (let t = x.MaxRetryWaitTime in t.TotalSeconds))
-            let c = Connector(x.Timeout, x.Retries, x.MaxRetryWaitTime, Log.Logger, mode=x.Mode)
-            c.Connect(appName, discovery)
+
+        member x.CreateConnector() =
+            let discovery = x.Discovery
+            Log.Information("Destination CosmosDb {mode} {endpointUri} timeout {timeout}s; Throttling retries {retries}, max wait {maxRetryWaitTime}s",
+                x.Mode, discovery.Endpoint, (let t = x.Timeout in t.TotalSeconds), x.Retries, (let t = x.MaxRetryWaitTime in t.TotalSeconds))
+            Equinox.CosmosStore.CosmosClientFactory(x.Timeout, x.Retries, x.MaxRetryWaitTime, mode=x.Mode)
+                .Connect(discovery)
+
+        member val Database =               a.TryGetResult Database   |> Option.defaultWith (fun () -> c.CosmosDatabase)
+        member val Container =              a.TryGetResult Container  |> Option.defaultWith (fun () -> c.CosmosContainer)
+        member x.Connect(connector) =
+            Log.Information("Destination CosmosDb Database {database} Container {container}", x.Database, x.Container)
+            Equinox.CosmosStore.CosmosStoreClient.Connect(connector, x.Database, x.Container)
 
     /// Parse the commandline; can throw exceptions in response to missing arguments and/or `-h`/`--help` args
     let parse tryGetConfigValue argv : Arguments =
@@ -174,34 +186,51 @@ module Args =
 
 let [<Literal>] AppName = "ArchiverTemplate"
 
+module CosmosStoreContext =
+
+    /// Create with default packing and querying policies. Search for other `module CosmosStoreContext` impls for custom variations
+    let create (storeClient : Equinox.CosmosStore.CosmosStoreClient) =
+        let maxEvents, maxJsonBytes = 100_000, 100_000 // default is 0, 30000
+        Equinox.CosmosStore.CosmosStoreContext(storeClient, tipMaxEvents=maxEvents, tipMaxJsonLength=maxJsonBytes)
+
 let build (args : Args.Arguments, log, storeLog : ILogger) =
     let source, (auxDiscovery, aux, leaseId, startFromTail, maxDocuments, lagFrequency) = args.SourceParams()
     let archiverSink =
         let target = source.Sink
-        let containers = Containers(target.Database, target.Container)
-        let conn = target.Connect AppName |> Async.RunSynchronously
-        let context = Equinox.Cosmos.Core.Context(conn, containers, storeLog)
-        CosmosSink.Start(log, args.MaxReadAhead, [|context|], args.MaxWriters, args.StatsInterval, args.StateInterval)
+        let connector = target.CreateConnector()
+        let context = target.Connect(connector) |> Async.RunSynchronously |> CosmosStoreContext.create
+        let eventsContext = Equinox.CosmosStore.Core.EventsContext(context, storeLog)
+        CosmosStoreSink.Start(log, args.MaxReadAhead, eventsContext, args.MaxWriters, args.StatsInterval, args.StateInterval, (*purgeInterval=TimeSpan.FromMinutes 10.,*) maxBytes = args.MaxBytes)
     let pipeline =
         let monitoredConnector, monitoredDiscovery, monitored  = source.MonitoringParams()
         let client, auxClient = monitoredConnector.CreateClient(AppName, monitoredDiscovery), monitoredConnector.CreateClient(AppName, auxDiscovery)
-        let createObserver () = CosmosSource.CreateObserver(log.ForContext<CosmosSource>(), archiverSink.StartIngester, Seq.collect Handler.selectArchivable)
-        CosmosSource.Run(log, client, monitored, aux,
+        let createObserver context = CosmosStoreSource.CreateObserver(log.ForContext<CosmosStoreSource>(), context, archiverSink.StartIngester, Seq.collect Handler.selectArchivable)
+        CosmosStoreSource.Run(log, client, monitored, aux,
             leaseId, startFromTail, createObserver,
             ?maxDocuments=maxDocuments, ?lagReportFreq=lagFrequency, auxClient=auxClient)
     archiverSink, pipeline
 
+// A typical app will likely have health checks etc, implying the wireup would be via `UseMetrics()` and thus not use this ugly code directly
+let startMetricsServer port : IDisposable =
+    let metricsServer = new Prometheus.KestrelMetricServer(port = port)
+    let ms = metricsServer.Start()
+    Log.Information("Prometheus /metrics endpoint on port {port}", port)
+    { new IDisposable with member x.Dispose() = ms.Stop(); (metricsServer :> IDisposable).Dispose() }
+
 let run args = async {
-    let log, storeLog = Log.ForContext<Propulsion.Streams.Scheduling.StreamSchedulingEngine>(), Log.ForContext<Equinox.Cosmos.Core.Context>()
+    let log, storeLog = Log.ForContext<Propulsion.Streams.Scheduling.StreamSchedulingEngine>(), Log.ForContext<Equinox.CosmosStore.Core.EventsContext>()
     let sink, pipeline = build (args, log, storeLog)
     pipeline |> Async.Start
-    return! sink.AwaitCompletion()
+    use _metricsServer : IDisposable = args.PrometheusPort |> Option.map startMetricsServer |> Option.toObj
+    do! sink.AwaitCompletion()
 }
 
 [<EntryPoint>]
 let main argv =
     try let args = Args.parse EnvVar.tryGet argv
-        try Log.Logger <- LoggerConfiguration().Configure(args.Verbose, args.SyncLogging, args.CfpVerbose).CreateLogger()
+        let appName = sprintf "archiver:%s" args.ConsumerGroupName
+        let appName = appName.Replace("ruben","")
+        try Log.Logger <- LoggerConfiguration().Configure(appName, args.ConsumerGroupName, args.Verbose, args.SyncLogging, args.CfpVerbose, args.MetricsEnabled).CreateLogger()
             try run args |> Async.RunSynchronously
                 0
             with e when not (e :? MissingArg) -> Log.Fatal(e, "Exiting"); 2