+++ CHANGELOG.md
@@ -15,6 +15,9 @@ The `Unreleased` section name is replaced by the expected version of next releas
 - `eqxProjector --source cosmos --kafka --synthesizeSequence`: Sample code for custom parsing of document changes [#84](https://github.com/jet/dotnet-templates/pull/84)
 - `feedApi`: Template for a Web API serving a feed that can be tracked by a `feedConsumer` [#88](https://github.com/jet/dotnet-templates/pull/88)
 - `feedConsumer`: Template for Feed Consumer tracking a `feedApi` [#88](https://github.com/jet/dotnet-templates/pull/88)
+- `proArchiver`: Parameterize max Sync batch size limit, default to 512KiB
+- `proAchiver`, `proPruner`, `eqxWeb`, `eqxWebCs`: Added `prometheus-net` integration
+- `proAchiver`, `proPruner`: Target `Equinox` v `3.0.0`, `Propulsion` v `2.10.0`
 
 ### Changed